Migrated codesegs 0 and 1 bss, migrated rodata for codeseg2_3.c, various matchings

+++ include/mus/n_libaudio_sc.h
@@ -45,8 +45,8 @@ typedef struct {
     ALLink      pFreeList;      /* list of free physical voices         */
     ALLink      pAllocList;     /* list of allocated physical voices    */
     ALLink      pLameList;      /* list of voices ready to be freed     */
-    s32         paramSamples;
-    s32         curSamples;     /* samples from start of game           */
+    s64         paramSamples;
+    s64         curSamples;     /* samples from start of game           */
     ALDMANew    dma;
     ALHeap      *heap;
     struct ALParam_s    *paramList;
